requirements.txt
appdirs==1.4.4
appnope==0.1.3
asgiref==3.6.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chart-studio==1.1.0
colorlover==0.3.0
contourpy==1.0.5
cryptography==39.0.1
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
Django==4.1.5
django-cors-headers==3.14.0
djangorestframework==3.14.0
entrypoints==0.4
executing==1.1.0
fonttools==4.37.4
frozendict==2.3.4
html5lib==1.1
idna==3.4
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.5.0
ipywidgets==8.0.4
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
multitasking==0.0.11
nest-asyncio==1.5.6
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.11.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==24.0.1
requests==2.28.1
retrying==1.3.4
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
statsmodels==0.13.2
tenacity==8.1.0
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.5
yfinance==0.2.11
